Stopped using the generic "unsigned int" in favor of more specific types.

+++ src/utils.h
@@ -286,7 +286,7 @@ bool logger_func(log_level_t level);
 
 #if defined(TRACING_ON)
 
-extern unsigned int _trace_level;
+extern uint8_t _trace_level;
 
 #define SHOW_TRACE(lvl) \
   (ledger::_log_level >= ledger::LOG_TRACE && lvl <= ledger::_trace_level)
